pyproject.toml
[tool.poetry]
name = "zebbeacon_discord"
version = "0.1.0"
description = ""
authors = ["Kevin <kevin@mancu.so>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
paho-mqtt = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
